feat: refactor login page to separate LoginForm component and implement session redirection
